+++ tomography/reconstruction.h
@@ -5,5 +5,5 @@
 #include "grid.h"
 using namespace std;
 
-void iterationArt(Grid &grid, const vector<VectorSparse> &a, const vector<float> &m, bool onlyPositive=true);
-void iterationSirt(Grid &grid, const vector<VectorSparse> &a, const vector<float> &m, bool onlyPositive=true);
+void iterationArt(Grid &grid, const vector<VectorSparse> &a, const vector<double> &m, bool onlyPositive=true);
+void iterationSirt(Grid &grid, const vector<VectorSparse> &a, const vector<double> &m, bool onlyPositive=true);
